Separated url data storing and accessing logic to another class
