+++ src/url/__init__.py
@@ -2,12 +2,12 @@ from litestar import Router
 
 #  Router "resolve" needs to be imported here to register it in root router separately
 from .routers import shorten, resolve  # noqa
-from .containers import Container
+from .containers import URLContainer
 from .settings import Settings
 
-def get_container() -> Container:
+def get_container() -> URLContainer:
     """Dependency container factory."""
-    container = Container()
+    container = URLContainer()
     settings = Settings()
 
     container.env.from_dict(settings.model_dump())
